01_matmul with EXLA on GPU
Cleaned up the EXLA notebook.  It demonstrates CPU BinaryBackend
and Nvidia GPU backend.  It shows how to switch between the backends
within a single notebook.

However, we couldn't quite demonstrate EXLA with CPU.  We tried to
show how to run EXLA with just the CPU.  However, we found a bug and
the Nx team immediately created a potential fix in master.  We couldn't
quickly get the workaround to function as expected.  We'll have to update
the notebook at some time in the future.